Move service and tests folder into Project
