+++ visualizer/src/Project/service/arraysMachine.js
