+++ visualizer/src/Project/service/imageMachine.test.js
