+++ visualizer/src/Project/service/raw/channelMachine.js
