+++ visualizer/src/Project/service/spotsMachine.js
