+++ visualizer/src/Project/service/tools/segment/brushMachine.js
