+++ visualizer/src/Project/service/tools/segment/thresholdMachine.js
