+++ visualizer/src/Project/service/tools/segment/toolUtils.js
