+++ visualizer/src/Project/service/tools/segment/trimMachine.js
