+++ visualizer/src/Project/service/tools/segmentMachine.js
